+++ src/main/java/io/zoemeow/dutapi/objects/accounts/ScheduleStudy.java
@@ -1,4 +1,4 @@
-package io.zoemeow.dutapi.objects;
+package io.zoemeow.dutapi.objects.accounts;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public class ScheduleStudy implements Serializable {
     private ArrayList<ScheduleItem> scheduleList = new ArrayList<>();
     private ArrayList<WeekItem> weekList = new ArrayList<>();
-    
+
     public ScheduleStudy() {
 
     }