+++ src/main/java/io/zoemeow/dutapi/objects/enums/LessonStatus.java
@@ -1,0 +1,7 @@
+package io.zoemeow.dutapi.objects.enums;
+
+public enum LessonStatus {
+    Unknown,
+    Leaving,
+    MakeUp
+}