+++ src/main/java/io/zoemeow/dutapi/objects/enums/NewsType.java
@@ -1,4 +1,4 @@
-package io.zoemeow.dutapi.objects;
+package io.zoemeow.dutapi.objects.enums;
 
 public enum NewsType {
     Global,